Renderscript request #2

+++ ImageSegmentation/src/com/tum/imagesegmentation/Constants.java
@@ -23,11 +23,11 @@ public final class Constants {
 	public static final int MAXPATH = 2;
 	public static final int NUMBER_SAMPLES = 2000;
 	public static final int NUMBER_SPLIT_LINES = 200;
-	public static final double ALPHA =3.0;
-	public static final double BETA = 1.0;
-	public static final double TAU = 0.16;
-	public static final double THETA = 0.5;
-	public static final double CONTOUR_THRESHOLD = 0.7;
+	public static final float ALPHA =3.0f;
+	public static final float BETA = 1.0f;
+	public static final float TAU = 0.16f;
+	public static final float THETA = 0.5f;
+	public static final float CONTOUR_THRESHOLD = 0.7f;
 	public static final double CONTOUR_WIDTH_RATIO = 0.01;
 	public static final int SCRIBBLE_RED_FG = 0;
 	public static final int SCRIBBLE_RED_BG = 255;
